aggregator/aggregate/requirements.txt
feedparser
isodate
mysql-connector
requests
XlsxWriter